+++ Ruby/ChangeLog.md
@@ -1,5 +1,11 @@
+## 3.0.1 (February 5, 2023)
+- Added configurable DNS SRV resolution
+- Added request type that attempts all SLP protocols
+- Improved debug messages
+
 ## 3.0.0 (January 31, 2023)
 - Added UT3/GS4 query support
+- Added debug mode
 - Refactored code
 - Added YARD documentation